refactor: PlayerName 값 객체 생성
